% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcontig_list-data.R
\docType{data}
\name{tcontig_list}
\alias{tcontig_list}
\title{Single chain immune repertoire dataset}
\format{
A list of two elements. First element ("data") is a list with data frames with clonotype tables.
Second element ("meta") is a metadata table.
\describe{
  \item{data}{List of immune repertoire data frames.}
  \item{meta}{Metadata}
  ...
}
}
\usage{
tcontig_list
}
\description{
A dataset with single chain TCR data for testing and examplatory purposes.
}
\concept{data}
\keyword{datasets}
